add skip-autogen=never to qemu module
This is needed to prevent jhbuild from skipping configure always:
qemu isn't a real autotools module so it already has some Makefile
in a clean git checkout. When jhbuild detects a Makefile in an
autogen sourcetree, it assumes it doesn't need to run configure.
skip-autogen forces it to always run configure even if a Makefile is
already present.